feat: reorganize exception classes into user package for better structure
