+++ backend/src/main/java/com/cinerama/backend/exception/user/PasswordsNotMatchException.java
@@ -1,4 +1,4 @@
-package com.cinerama.backend.exception;
+package com.cinerama.backend.exception.user;
 
 public class PasswordsNotMatchException extends RuntimeException {
     public PasswordsNotMatchException(String message) {
