+++ backend/src/main/java/com/cinerama/backend/exception/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.cinerama.backend.exception;
+package com.cinerama.backend.exception.user;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException() {
